add: architecture document
